+++ FitPlannerWidgets/Small WIdget/SmallWIdgetEntry.swift
@@ -1,0 +1,13 @@
+//
+//  SmallWIdgetEntry.swift
+//  FitPlanner
+//
+//  Created by Alberto Dominguez on 1/12/22.
+//
+
+import WidgetKit
+import SwiftUI
+
+struct SmallWidgetEntry: TimelineEntry {
+    let date: Date
+}
